processing_service/requirements.txt
redis==5.0.1
rq==1.15.1
pypdf==3.17.4
python-docx==1.1.0
langchain==0.1.0
langchain-text-splitters==0.0.1
langchain-community==0.0.13
pydantic==2.5.0
pydantic-settings==2.1.0